@@ -19,8 +19,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     testOptions {
         unitTests.returnDefaultValues = true
@@ -28,6 +28,9 @@ android {
     productFlavors {
     }
     namespace 'com.ramotion.foldingcell'
+   /* kotlinOptions {
+        jvmTarget = '1.8'
+    }*/
 }
 
 dependencies {
