@@ -27,6 +27,7 @@ import com.journeyapps.barcodescanner.BarcodeEncoder;
 import com.safesoft.proapp.distribute.MainActivity;
 import com.safesoft.proapp.distribute.R;
 import com.safesoft.proapp.distribute.activities.splashscreen.splashScreen;
+import com.safesoft.proapp.distribute.app.BaseApplication;
 
 import de.keyboardsurfer.android.widget.crouton.Crouton;
 import de.keyboardsurfer.android.widget.crouton.Style;
@@ -42,26 +43,8 @@ public class ActivityActivation extends AppCompatActivity {
     EditText code_activation_input;
     public static final String PREFS = "ALL_PREFS";
 
-    //////////////ancien declaration ////////////
-    private static final String TAG = ActivityActivation.class.getSimpleName();
-
-    private static final int READ_PHONE_STATE = 5;
-    private MediaPlayer mp;
-    //private String PREFS = "ALL_PREFS";
-
-    private ImageButton btnActiver;
-    private ProgressDialog progressDialog;
-    private final boolean state_a = false;
-    private final boolean phone_read_state = false;
-
-    private Animation animation;
-    private Animation animation1;
-
-    TelephonyManager telephonyManager;
-
-    private EasyTextInputLayout _code_client, _nom_client, _numero_tele, _revendeur;
-
-    String deviceID = "";
+    public ActivityActivation() {
+    }
     //////////////ancien declaration ////////////
 
     @Override
@@ -69,8 +52,6 @@ public class ActivityActivation extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         getSupportActionBar().hide();
         setContentView(R.layout.activity_activation);
-
-        //telephonyManager = (TelephonyManager) getSystemService(ActivityActivation.TELEPHONY_SERVICE);
 
         inforevenedeur1 = findViewById(R.id.InfoRevendeur1);
         inforevenedeur2 = findViewById(R.id.InfoRevendeur2);
@@ -177,6 +158,14 @@ public class ActivityActivation extends AppCompatActivity {
 ///////////////////////////////////  old activation//////////////
 
     }
+
+    public void saveActivationData(boolean is_activated) {
+        SharedPreferences sharedPreferences = getSharedPreferences(PREFS,MODE_PRIVATE);
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        editor.putBoolean("APP_ACTIVATED", is_activated);
+        editor.apply();
+    }
+
     public void onClick_Btn_Activation(View view) {
         Button btn_start = findViewById(R.id.BTN_START);
         code_activation_input = findViewById(R.id.CODE_ACTIVATION);
@@ -201,11 +190,14 @@ public class ActivityActivation extends AppCompatActivity {
             Crouton.makeText(ActivityActivation.this, "Code d'activation correct", Style.INFO).show();
             btn_start.setVisibility(View.VISIBLE);  ///// 357272286 ///// 350889357
             saveData();
+            saveActivationData(true);
+
         } else {
             //Toast.makeText(ActivityActivation.this, "Code d'activation incorrect",Toast.LENGTH_SHORT).show();
             Crouton.makeText(ActivityActivation.this, "Code d'activation incorrect", Style.ALERT).show();
             ((EditText)findViewById(R.id.CODE_ACTIVATION)).setText("");
             btn_start.setVisibility(View.INVISIBLE);
+            saveActivationData(false);
         }
     }
 
@@ -214,263 +206,6 @@ public class ActivityActivation extends AppCompatActivity {
         overridePendingTransition(R.anim.slide_in_right, R.anim.slide_out_left);
     }
 
-
-//    @Override
-//    protected void onStart() {
-//        animation = AnimationUtils.loadAnimation(this, R.anim.slide_left_button_activation);
-//        animation1 = AnimationUtils.loadAnimation(this, R.anim.slide_right_button_activation);
-//
-//        startAnimation();
-//
-//        super.onStart();
-//    }
-
-//    protected void startAnimation() {
-//        btnActiver.startAnimation(animation);
-////        notice.startAnimation(animation1);
-//    }
-
-//    @Override
-//    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-//        if (requestCode == READ_PHONE_STATE) {
-//            if (ActivityCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE) != PackageManager.PERMISSION_GRANTED) {
-//                return;
-//            }
-//            phone_read_state = true;
-//            getSupportActionBar().setSubtitle("Activation ( MEI : " + deviceID + " )");
-//            sendRequest("gg");
-//        }
-//        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-//    }
-
-//    public void onClickButton(View v) {
-//        switch (v.getId()) {
-//            case R.id.activer:
-//                if (!state_a) {
-//                    if (check_fields()) {
-//                        sendRequest("gg");
-//                    }
-//                } else {
-//                    startActivity(new Intent(getApplicationContext(), splashScreen.class));
-//
-//
-//                }
-//
-//                break;
-//        }
-//    }
-
-//    protected boolean check_fields() {
-//        boolean check = true;
-//
-//        if (_code_client.getEditText().getText().length() <= 0) {
-//            check = false;
-//            _code_client.getEditText().setError("Code client est obligatoire");
-//        } else if (_code_client.getEditText().getText().length() != 6) {
-//            if (_code_client.getEditText().getText().length() == 1) {
-//                if (!_code_client.getEditText().getText().toString().equals("0")) {
-//                    check = false;
-//                    _code_client.getEditText().setError("Le code client doit contient que 6 digits ou la valuer 0");
-//                }
-//            } else {
-//                check = false;
-//                _code_client.getEditText().setError("Le code client doit contient que 6 digits ou la valuer 0");
-//            }
-//
-//        }
-//
-//        if (_nom_client.getEditText().getText().length() <= 0) {
-//            check = false;
-//            _nom_client.getEditText().setError("Nom client est obligatoire");
-//        }
-//
-//        if (_numero_tele.getEditText().getText().length() <= 0) {
-//            check = false;
-//            _numero_tele.getEditText().setError("Numéro du téléphone est obligatoire");
-//        }
-//
-//        /*
-//        if(_revendeur.getEditText().getText().length() <= 0){
-//            check = false;
-//            _revendeur.getEditText().setError("Revendeur est obligatoire");
-//        }
-//        */
-//
-//        return check;
-//    }
-
-//    public void sendRequest(String request) {
-//        try {
-//            if (request.length() > 0) {
-//
-//                progressDialog = ProgressDialog.show(ActivityActivation.this, getString(R.string.leading_activation), getString(R.string.activation_wait));
-//
-//                new Thread(new Runnable() {
-//                    @Override
-//                    public void run() {
-//
-//                        if (hasActiveInternetConnection()) {
-//                            try {
-//                                NetClient nc = new NetClient("105.96.9.62", 2578); // ip adress and port
-//
-//                                if (ActivityCompat.checkSelfPermission(ActivityActivation.this, Manifest.permission.READ_PHONE_STATE) != PackageManager.PERMISSION_GRANTED) {
-//                                    /*runOnUiThread(new Runnable() {
-//                                        @Override
-//                                        public void run() {
-//                                            if (progressDialog != null && !ActivityActivation.this.isFinishing()) {
-//                                                progressDialog.dismiss();
-//                                                progressDialog = null;
-//                                            }
-//                                            Sound(R.raw.back);
-//                                            Crouton.makeText(ActivityActivation.this, "Permission de lire IMEI de téléphone n'est encore garanti",Style.ALERT).show();
-//                                        }
-//                                    });*/
-//                                    return;
-//                                }
-//                                String message_online = "DISTRIBUTION." + deviceID + "." + _code_client.getEditText().getText() + "." + _nom_client.getEditText().getText() + "." + _numero_tele.getEditText().getText() + ".SAFESOFT";
-//                                nc.sendDataWithString(message_online);
-//                                Log.d("TAG", "run: " + message_online);
-//                                final String r = nc.receiveDataFromServer();
-//                                Log.d("TAG", "run: " + r);
-//                                if (r.equals("NON")) {
-//                                    runOnUiThread(new Runnable() {
-//                                        @Override
-//                                        public void run() {
-//                                            if (progressDialog != null && !ActivityActivation.this.isFinishing()) {
-//                                                progressDialog.dismiss();
-//                                                progressDialog = null;
-//                                            }
-//                                            Sound(R.raw.back);
-//                                            Crouton.makeText(ActivityActivation.this, "Demande envoyée", Style.INFO).show();
-//                                        }
-//                                    });
-//                                } else {
-//                                    Sound(R.raw.beep);
-//                                    runOnUiThread(new Runnable() {
-//                                        @Override
-//                                        public void run() {
-//                                            String token = "";
-//                                            String nom_client = "";
-//                                            String telephone = "";
-//                                            token = r.substring(0, r.indexOf(";"));
-//                                            if (token.equals("OK")) {
-//                                                nom_client = r.substring(r.indexOf(";") + 1, r.lastIndexOf(";"));
-//                                                telephone = r.substring(r.lastIndexOf(";") + 1, r.length());
-//                                                Crouton.makeText(ActivityActivation.this, " Opération success \n pour le client  : " + nom_client + " / " + telephone, Style.CONFIRM).show();
-//                                                getSupportActionBar().setTitle("Distribute ( Version enregistrée )");
-//
-//                                                SharedPreferences.Editor editor = getSharedPreferences(PREFS_ACTIVATION, MODE_PRIVATE).edit();
-//                                                editor.putString("NOM_CLIENT", nom_client.toString());
-//                                                editor.putString("TELEPHONE", telephone.toString());
-//                                                editor.putBoolean("IS_ACTIVATED", true);
-//                                                // editor.putString("MEI", activity.getText().toString());
-//                                                editor.commit();
-//                                                state_a = true;
-//
-//                                                btnActiver.setImageResource(R.mipmap.commencer);
-//
-//
-//                                                _code_client.getEditText().setEnabled(false);
-//                                                _nom_client.getEditText().setEnabled(false);
-//                                                _numero_tele.getEditText().setEnabled(false);
-//                                                _revendeur.getEditText().setEnabled(false);
-//
-//                                                startAnimation();
-//                                            } else {
-//                                                Crouton.makeText(ActivityActivation.this, "Erreur dans la réponse !", Style.CONFIRM).show();
-//                                            }
-//
-//                                            if (progressDialog != null && !ActivityActivation.this.isFinishing()) {
-//                                                progressDialog.dismiss();
-//                                                progressDialog = null;
-//                                            }
-//                                        }
-//                                    });
-//                                }
-//
-//                                Log.e("TRACKKK", "==========> " + r);
-//
-//                            } catch (Exception e) {
-//                                Log.e("TRACKKK", "Server is down ");
-//                                Log.e("TRACKKK", e.getMessage());
-//
-//                                runOnUiThread(new Runnable() {
-//                                    @Override
-//                                    public void run() {
-//                                        if (progressDialog != null && !ActivityActivation.this.isFinishing()) {
-//                                            progressDialog.dismiss();
-//                                            progressDialog = null;
-//                                        }
-//                                        Sound(R.raw.back);
-//                                        Crouton.makeText(ActivityActivation.this, "Problème de Connexion Internt ou Réseau mobile", Style.ALERT).show();
-//                                    }
-//                                });
-//                            }
-//                        } else {
-//                            runOnUiThread(new Runnable() {
-//                                @Override
-//                                public void run() {
-//                                    if (progressDialog != null && !ActivityActivation.this.isFinishing()) {
-//                                        progressDialog.dismiss();
-//                                        progressDialog = null;
-//                                    }
-//                                    Sound(R.raw.back);
-//                                    Crouton.makeText(ActivityActivation.this, "Problème de Connexion Internt ou Réseau mobile", Style.ALERT).show();
-//                                }
-//                            });
-//                        }
-//                    }
-//                }).start();
-//
-//            } else {
-//                //On affiche un message d'erreur dans un Toast
-//                Crouton.makeText(ActivityActivation.this, "Demande non envoyé, requete vide !", Style.ALERT).show();
-//            }
-//        } catch (Exception e) {
-//            Log.e("TRACKKK", "Error checking internet connection", e);
-//            Toast.makeText(this, "Problème dans l'envoie de la demande!", Toast.LENGTH_SHORT).show();
-//        } finally {
-//
-//        }
-//    }
-
-//    public boolean hasActiveInternetConnection() {
-//        return true;
-//    }
-//
-//    public void Sound(int SourceSound) {
-//        mp = MediaPlayer.create(this, SourceSound);
-//        mp.start();
-//    }
-//
-//    public void requestPermission() {
-//
-//        /*if (ContextCompat.checkSelfPermission(this, Manifest.permission.RECEIVE_SMS) != PackageManager.PERMISSION_GRANTED) {
-//            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.RECEIVE_SMS}, RECIEVE_SMS);
-//        }
-//
-//        if (ContextCompat.checkSelfPermission(this, Manifest.permission.SEND_SMS) != PackageManager.PERMISSION_GRANTED) {
-//            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.SEND_SMS}, SEND_SMS_PERMISSION);
-//        }
-//
-//        if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-//            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.ACCESS_FINE_LOCATION}, ACCES_FINE_LOCATION);
-//            checkPermission = false;
-//        }else{
-//            checkPermission = true;
-//        }
-//
-//        if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-//            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA}, ZBAR_CAMERA_PERMISSION);
-//        }
-//*/
-//        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE) != PackageManager.PERMISSION_GRANTED) {
-//            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.READ_PHONE_STATE}, READ_PHONE_STATE);
-//            phone_read_state = false;
-//        } else {
-//            phone_read_state = true;
-//        }
-//    }
     public void saveData() {
         SharedPreferences sharedPreferences = getSharedPreferences(PREFS,MODE_PRIVATE);
         SharedPreferences.Editor editor = sharedPreferences.edit();
