@@ -41,17 +41,19 @@ import com.google.android.gms.vision.barcode.Barcode;
 import com.safesoft.proapp.distribute.R;
 import com.safesoft.proapp.distribute.activities.ActivityHtmlView;
 import com.safesoft.proapp.distribute.activities.pdf.GeneratePDF;
-import com.safesoft.proapp.distribute.adapters.ListViewAdapterPanier;
+import com.safesoft.proapp.distribute.adapters.ListViewAdapterPanierAchat;
+import com.safesoft.proapp.distribute.adapters.ListViewAdapterPanierVente;
 import com.safesoft.proapp.distribute.adapters.RecyclerAdapterCheckProducts;
 import com.safesoft.proapp.distribute.databases.DATABASE;
 import com.safesoft.proapp.distribute.eventsClasses.ByteDataEvent;
+import com.safesoft.proapp.distribute.eventsClasses.CheckedPanierEventAchat2;
 import com.safesoft.proapp.distribute.eventsClasses.CheckedPanierEventBon2;
 import com.safesoft.proapp.distribute.eventsClasses.LocationEvent;
 import com.safesoft.proapp.distribute.eventsClasses.RemiseEvent;
 import com.safesoft.proapp.distribute.eventsClasses.SelectedFournisseurEvent;
 import com.safesoft.proapp.distribute.eventsClasses.TimbreEvent;
 import com.safesoft.proapp.distribute.eventsClasses.ValidateFactureEvent;
-import com.safesoft.proapp.distribute.fragments.FragmentQte;
+import com.safesoft.proapp.distribute.fragments.FragmentQteAchat;
 import com.safesoft.proapp.distribute.fragments.FragmentRemise;
 import com.safesoft.proapp.distribute.fragments.FragmentSelectFournisseur;
 import com.safesoft.proapp.distribute.fragments.FragmentSelectProduct;
@@ -59,10 +61,12 @@ import com.safesoft.proapp.distribute.fragments.FragmentTimbre;
 import com.safesoft.proapp.distribute.fragments.FragmentValideBon;
 import com.safesoft.proapp.distribute.gps.ServiceLocation;
 import com.safesoft.proapp.distribute.postData.PostData_Achat1;
+import com.safesoft.proapp.distribute.postData.PostData_Achat2;
 import com.safesoft.proapp.distribute.postData.PostData_Bon2;
 import com.safesoft.proapp.distribute.postData.PostData_Fournisseur;
 import com.safesoft.proapp.distribute.postData.PostData_Produit;
 import com.safesoft.proapp.distribute.printing.Printing;
+import com.safesoft.proapp.distribute.utils.Env;
 
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
@@ -92,10 +96,10 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
     private static final int ACCES_FINE_LOCATION = 2;
     private static final int CAMERA_PERMISSION = 5;
     private Intent intent_location;
-    private ListViewAdapterPanier PanierAdapter;
+    private ListViewAdapterPanierAchat PanierAdapter;
     private Button btn_select_fournisseur;
     private DATABASE controller;
-    private  ArrayList<PostData_Bon2> final_panier;
+    private  ArrayList<PostData_Achat2> final_panier;
     private TextView total_ht, tva, txv_timbre, txv_remise, total_ttc, total_ttc_remise;
     private double val_total_ht = 0.00;
     private double val_tva = 0.00;
@@ -120,7 +124,6 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
     SharedPreferences prefs;
 
     // constant code for runtime permissions
-    private static final int PERMISSION_REQUEST_CODE = 200;
 
     @SuppressLint("SimpleDateFormat") SimpleDateFormat date_format;
     @SuppressLint("SimpleDateFormat") SimpleDateFormat heure_format;
@@ -239,10 +242,11 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                     "ACHAT2.COLISSAGE, " +
                     "ACHAT2.QTE, " +
                     "ACHAT2.QTE_GRAT, " +
-                    "ACHAT2.PU, " +
+                    "ACHAT2.PA_HT, " +
                     "ACHAT2.TVA, " +
                     "ACHAT2.CODE_DEPOT, " +
                     "ACHAT2.QTE_GRAT, " +
+                    "PRODUIT.PAMP, " +
                     "PRODUIT.STOCK " +
                     "FROM ACHAT2 " +
                     "LEFT JOIN PRODUIT ON (ACHAT2.CODE_BARRE = PRODUIT.CODE_BARRE) " +
@@ -256,7 +260,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
             onFournisseurSelected(fournisseur_selected);
 
             // Create the adapter to convert the array to views
-            PanierAdapter = new ListViewAdapterPanier(this, R.layout.transfert2_items, final_panier, TYPE_ACTIVITY);
+            PanierAdapter = new ListViewAdapterPanierAchat(this, R.layout.transfert2_items, final_panier, TYPE_ACTIVITY);
 
             expandableListView = findViewById(R.id.expandable_listview);
 
@@ -372,9 +376,18 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                         .show();
                 return;
                 }
-                if( achat1.fournis.length() <1){
+                if(achat1.fournis.isEmpty()){
 
                     Crouton.makeText(ActivityAchat.this, "Vous devez Séléctionner un fournisseur tout d'abord", Style.ALERT).show();
+                    return;
+                }
+
+                if(!prefs.getBoolean("APP_ACTIVATED",false) && final_panier.size() >= 2){
+                    new SweetAlertDialog(ActivityAchat.this, SweetAlertDialog.ERROR_TYPE)
+                            .setTitleText("Important !")
+                            .setContentText(Env.MESSAGE_DEMANDE_ACTIVITATION)
+                            .show();
+
                     return;
                 }
 
@@ -387,7 +400,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                 break;
 
             case R.id.valide_facture:
-                if( final_panier.size() <1){
+                if(final_panier.isEmpty()){
                     Crouton.makeText(ActivityAchat.this, "Votre panier est vide !", Style.ALERT).show();
                     return;
                 }
@@ -451,7 +464,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                                 .show();
                         return;
                     }
-                    if( achat1.fournis.length() <1){
+                    if(achat1.fournis.isEmpty()){
 
                         Crouton.makeText(ActivityAchat.this, "Vous devez Séléctionner un client tout d'abord", Style.ALERT).show();
                         return;
@@ -523,7 +536,8 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                     bactivity = ActivityAchat.this;
 
                     Printing printer = new Printing();
-                    printer.start_print_bon(bactivity, "ACHAT", final_panier, null, achat1);
+                    printer.start_print_bon_achat(bactivity, "ACHAT", final_panier, achat1);
+
                 }else{
                     Intent html_intent = new Intent(this, ActivityHtmlView.class);
                     html_intent.putExtra("TYPE_BON" , "ACHAT");
@@ -595,16 +609,17 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                 "ACHAT2.COLISSAGE, " +
                 "ACHAT2.QTE, " +
                 "ACHAT2.QTE_GRAT, " +
-                "ACHAT2.PU, " +
+                "ACHAT2.PA_HT, " +
                 "ACHAT2.TVA, " +
                 "ACHAT2.CODE_DEPOT, " +
+                "PRODUIT.PAMP, " +
                 "PRODUIT.STOCK " +
                 "FROM ACHAT2 " +
                 "LEFT JOIN PRODUIT ON (ACHAT2.CODE_BARRE = PRODUIT.CODE_BARRE) " +
                 "WHERE ACHAT2.NUM_BON = '" + NUM_BON + "'" );
 
         // Create the adapter to convert the array to views
-        PanierAdapter = new ListViewAdapterPanier(this, R.layout.transfert2_items, final_panier, TYPE_ACTIVITY);
+        PanierAdapter = new ListViewAdapterPanierAchat(this, R.layout.transfert2_items, final_panier, TYPE_ACTIVITY);
 
         expandableListView = findViewById(R.id.expandable_listview);
 
@@ -654,9 +669,17 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                                 SOURCE = "ACHAT2_DELETE";
                                 assert info != null;
                                 controller.delete_from_achat2("ACHAT2", final_panier.get(info.position).recordid, final_panier.get(info.position));
+
+                                controller.update_pamp_on_delete_produit(final_panier.get(info.position).codebarre,
+                                        final_panier.get(info.position).stock_produit,
+                                        final_panier.get(info.position).pa_ht_produit,
+                                        final_panier.get(info.position).qte,
+                                        final_panier.get(info.position).pa_ht,
+                                        final_panier.get(info.position).pa_ht);
+
                                 initData();
                                 //PanierAdapter.RefrechPanier(final_panier);
-                                PanierAdapter = new ListViewAdapterPanier(ActivityAchat.this, R.layout.transfert2_items, final_panier, TYPE_ACTIVITY);
+                                PanierAdapter = new ListViewAdapterPanierAchat(ActivityAchat.this, R.layout.transfert2_items, final_panier, TYPE_ACTIVITY);
                                 expandableListView.setAdapter(PanierAdapter);
 
                             } catch (Exception e) {
@@ -683,8 +706,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                     SOURCE = "ACHAT2_EDIT";
                     Activity activity;
                     activity = ActivityAchat.this;
-                    //double last_price = controller.select_last_price_from_database(achat1.code_frs, final_panier.get(info.position).codebarre);
-                    FragmentQte fragmentqte = new FragmentQte();
+                    FragmentQteAchat fragmentqte = new FragmentQteAchat();
                     assert info != null;
                     fragmentqte.showDialogbox(SOURCE, activity, getBaseContext(), final_panier.get(info.position), 0);
 
@@ -714,7 +736,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
 
         for(int k = 0; k< final_panier.size(); k++){
 
-            double total_montant_produit = final_panier.get(k).p_u * final_panier.get(k).qte;
+            double total_montant_produit = final_panier.get(k).pa_ht * final_panier.get(k).qte;
             double montant_tva_produit = total_montant_produit  * ((final_panier.get(k).tva) / 100);
             val_total_ht = val_total_ht + total_montant_produit;
             val_tva = val_tva + montant_tva_produit;
@@ -823,7 +845,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
         achat1.fournis = fournisseur_selected.code_frs;
         //bon1_a_com.code_vendeur = "000000";
         //bon1_temp.mode_tarif = client_selected.mode_tarif;
-        //achat1.solde_ancien = fournisseur_selected.solde_montant;
+        achat1.solde_ancien = fournisseur_selected.solde_montant;
         achat1.nbr_p = final_panier.size();
 
 
@@ -879,9 +901,6 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
 
 
     private void startScanProduct() {
-        /**
-         * Build a new MaterialBarcodeScanner
-         */
 
         final MaterialBarcodeScanner materialBarcodeScanner = new MaterialBarcodeScannerBuilder()
                 .withActivity(ActivityAchat.this)
@@ -915,50 +934,49 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
     @Override
     public void onClick(View v, int position, PostData_Produit item) throws ParseException {
 
-            PostData_Bon2 bon2 = new PostData_Bon2();
-            bon2.produit = item.produit;
-            bon2.codebarre = item.code_barre;
-            bon2.stock_produit = item.stock;
-            bon2.destock_type = item.destock_type;
-            bon2.destock_code_barre = item.destock_code_barre;
-            bon2.destock_qte = item.destock_qte;
-            bon2.tva = item.tva;
-            bon2.colissage = item.colissage;
-            bon2.p_u = item.pa_ht;
-            bon2.promo = item.promo;
-            bon2.d1 = item.d1;
-            bon2.d2 = item.d2;
-            bon2.pp1_ht = item.pp1_ht;
+        PostData_Achat2 achat2 = new PostData_Achat2();
+        achat2.produit = item.produit;
+        achat2.codebarre = item.code_barre;
+        achat2.stock_produit = item.stock;
+        achat2.tva = item.tva;
+        achat2.colissage = item.colissage;
+        achat2.pa_ht = item.pa_ht;
+        achat2.pa_ht_produit = item.pa_ht;
+        achat2.num_bon = NUM_BON;
+        achat2.code_depot = CODE_DEPOT;
 
-            bon2.num_bon = NUM_BON;
-            bon2.code_depot = CODE_DEPOT;
-            SOURCE = "ACHAT2_INSERT";
-            Activity activity = ActivityAchat.this;
-            //double last_price = controller.select_last_price_from_database(achat1.code_frs, final_panier.get(position).codebarre);
-            FragmentQte fragmentqte = new FragmentQte();
-            fragmentqte.showDialogbox(SOURCE, activity, getBaseContext(),  bon2, 0);
+        SOURCE = "ACHAT2_INSERT";
+
+        Activity activity = ActivityAchat.this;
+        //double last_price = controller.select_last_price_from_database(achat1.code_frs, final_panier.get(position).codebarre);
+        FragmentQteAchat fragmentqte = new FragmentQteAchat();
+        fragmentqte.showDialogbox(SOURCE, activity, getBaseContext(),  achat2, 0);
 
         //Save clicked item position in list
         //save permanently
         prefs = getSharedPreferences(PREFS, MODE_PRIVATE);
         prefs.edit().putInt("LAST_CLICKED_POSITION", position).apply();
+
     }
 
     @Subscribe
-    public void onItemPanierReceive(CheckedPanierEventBon2 item_panier){
-
+    public void onItemPanierReceive(CheckedPanierEventAchat2 item_panier){
            try {
+
+               controller.update_pamp_on_insert_update_produit(item_panier.getData().codebarre, item_panier.getData().stock_produit,item_panier.getData().pa_ht_produit, item_panier.getData().qte + item_panier.getData().gratuit, item_panier.getData().pa_ht, item_panier.getData().pa_ht);
+
+
                if(SOURCE.equals("ACHAT2_INSERT")){
                    if(item_panier.getIfExist()){
                        controller.update_into_achat2("ACHAT2", NUM_BON, item_panier.getData(), item_panier.getQteOld(),item_panier.getGratuitOld());
+                       //update pamp produit
+
                    }else {
                        controller.insert_into_achat2("ACHAT2",  item_panier.getData());
                    }
-
                }else if(SOURCE.equals("ACHAT2_EDIT")){
                    controller.update_into_achat2("ACHAT2", NUM_BON, item_panier.getData(), item_panier.getQteOld(),item_panier.getGratuitOld());
                }
-
                initData();
                Sound(R.raw.cashier_quotka);
 
@@ -984,13 +1002,18 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
         achat1.date_f = date_format.format(c.getTime());
         achat1.heure_f = heure_format.format(c.getTime());
 
+
         //bon1.diff_time =
         if (controller.validate_achat1_sql("ACHAT1", achat1)) {
             achat1.blocage = "F";
         }
 
         validate_theme();
+
+
     }
+
+
     public void validate_theme(){
         if (achat1.blocage.equals("F")) {
             //findViewById(R.id.client).setBackgroundColor(Color.LTGRAY);
@@ -1010,7 +1033,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
 
     private void selectProductFromScan(String resultscan) throws ParseException {
         ArrayList<PostData_Produit> produits;
-        PostData_Bon2 bon2 = new PostData_Bon2();
+        PostData_Achat2 achat2 = new PostData_Achat2();
 
             String querry = "SELECT PRODUIT_ID, CODE_BARRE, REF_PRODUIT, PRODUIT, PA_HT, TVA, PAMP, PROMO, D1, D2, PP1_HT, PV1_HT, PV2_HT, PV3_HT, PV4_HT, PV5_HT, PV6_HT, STOCK, COLISSAGE, PHOTO, DETAILLE, ISNEW, FAMILLE, DESTOCK_TYPE, " +
                     "CASE WHEN PRODUIT.COLISSAGE <> 0 THEN  (PRODUIT.STOCK/PRODUIT.COLISSAGE) ELSE 0 END STOCK_COLIS , DESTOCK_CODE_BARRE," +
@@ -1018,7 +1041,7 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
                     "FROM PRODUIT  WHERE CODE_BARRE = '" + resultscan + "' OR REF_PRODUIT = '" + resultscan + "'";
             produits = controller.select_produits_from_database(querry);
 
-            if(produits.size() == 0){
+            if(produits.isEmpty()){
                 String querry1 = "SELECT * FROM CODEBARRE WHERE CODE_BARRE_SYN = '"+resultscan+"'";
                 String code_barre = controller.select_codebarre_from_database(querry1);
 
@@ -1030,27 +1053,21 @@ public class ActivityAchat extends AppCompatActivity implements RecyclerAdapterC
             }
 
         if(produits.size() == 1){
-            bon2.num_bon = NUM_BON;
-            bon2.code_depot = CODE_DEPOT;
-            bon2.produit = produits.get(0).produit;
-            bon2.codebarre = produits.get(0).code_barre;
-            bon2.stock_produit = produits.get(0).stock;
-            bon2.destock_type = produits.get(0).destock_type;
-            bon2.destock_code_barre = produits.get(0).destock_code_barre;
-            bon2.destock_qte = produits.get(0).destock_qte;
-            bon2.tva = produits.get(0).tva;
-            bon2.colissage = produits.get(0).colissage;
-            bon2.p_u = produits.get(0).pa_ht;
-            bon2.promo = produits.get(0).promo;
-            bon2.d1 = produits.get(0).d1;
-            bon2.d2 = produits.get(0).d2;
-            bon2.pp1_ht = produits.get(0).pp1_ht;
+            achat2.num_bon = NUM_BON;
+            achat2.code_depot = CODE_DEPOT;
+            achat2.produit = produits.get(0).produit;
+            achat2.codebarre = produits.get(0).code_barre;
+            achat2.stock_produit = produits.get(0).stock;
+            achat2.tva = produits.get(0).tva;
+            achat2.colissage = produits.get(0).colissage;
+            achat2.pa_ht = produits.get(0).pa_ht;
+            achat2.pa_ht_produit = produits.get(0).pa_ht;
 
 
             SOURCE = "ACHAT2_INSERT";
             Activity activity = ActivityAchat.this;
-            FragmentQte fragmentqte = new FragmentQte();
-            fragmentqte.showDialogbox(SOURCE, activity, getBaseContext(),  bon2, 0);
+            FragmentQteAchat fragmentqte = new FragmentQteAchat();
+            fragmentqte.showDialogbox(SOURCE, activity, getBaseContext(),  achat2, 0);
 
         }else if(produits.size() > 1){
             Crouton.makeText(ActivityAchat.this, "Attention il y a 2 produits avec le meme code !", Style.ALERT).show();
