@@ -1,7 +1,6 @@
 package com.safesoft.proapp.distribute.fragments;
 
 import static android.content.Context.MODE_PRIVATE;
-import static androidx.core.app.ActivityCompat.startActivityForResult;
 import static com.rilixtech.materialfancybutton.MaterialFancyButton.POSITION_LEFT;
 
 import android.app.Activity;
@@ -33,6 +32,7 @@ import com.safesoft.proapp.distribute.R;
 import com.safesoft.proapp.distribute.databases.DATABASE;
 import com.safesoft.proapp.distribute.eventsClasses.ByteDataEvent;
 import com.safesoft.proapp.distribute.eventsClasses.ProductEvent;
+import com.safesoft.proapp.distribute.eventsClasses.SelectedClientEvent;
 import com.safesoft.proapp.distribute.postData.PostData_Params;
 import com.safesoft.proapp.distribute.postData.PostData_Produit;
 
@@ -48,7 +48,7 @@ import java.util.Random;
 import de.keyboardsurfer.android.widget.crouton.Crouton;
 import de.keyboardsurfer.android.widget.crouton.Style;
 
-public class FragmentNewProduct {
+public class FragmentNewEditProduct {
 
     final String ALLOWED_CHARACTERS_CODEBARRE ="0123456789ABCDEFGHJK";
     final String ALLOWED_CHARACTERS_REFERENCE ="012345678-9RSTUVWXYZ";
@@ -84,27 +84,28 @@ public class FragmentNewProduct {
             txt_input_prix5_ht, txt_input_prix5_ttc,
             txt_input_prix6_ht, txt_input_prix6_ttc;
     LinearLayout ly_prix_achat;
-
     EventBus bus = EventBus.getDefault();
     Activity activity;
     AlertDialog dialog;
     ImageView img_product;
     byte[] inputData = null;
-
     private final String PREFS = "ALL_PREFS";
 
     PostData_Produit created_produit;
+    PostData_Produit old_product;
     private DATABASE controller;
     NumberFormat nf,nq;
     private Barcode barcodeResult;
 
     private PostData_Params params;
+    SharedPreferences prefs;
     //PopupWindow display method
 
-    public void showDialogbox(Activity activity, String SOURCE_ACTIVITY) {
+    public void showDialogbox(Activity activity, String SOURCE_ACTIVITY, PostData_Produit old_product) {
 
         this.activity = activity;
         this.controller = new DATABASE(activity);
+        this.old_product = old_product;
 
         // Declare US print format
         nf = NumberFormat.getInstance(Locale.US);
@@ -113,7 +114,7 @@ public class FragmentNewProduct {
         nq = NumberFormat.getInstance(Locale.US);
         ((DecimalFormat) nq).applyPattern("####0.##");
 
-        SharedPreferences prefs = this.activity.getSharedPreferences(PREFS, MODE_PRIVATE);
+        prefs = activity.getSharedPreferences(PREFS, MODE_PRIVATE);
         created_produit = new PostData_Produit();
 
         AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(activity);
@@ -133,7 +134,7 @@ public class FragmentNewProduct {
         dialog.getWindow().setAttributes(layoutParams);
 
 
-        btn_valider = (MaterialFancyButton) dialogview.findViewById(R.id.btn_remise);
+        btn_valider = dialogview.findViewById(R.id.btn_remise);
         btn_valider.setBackgroundColor(Color.parseColor("#3498db"));
         btn_valider.setFocusBackgroundColor(Color.parseColor("#5474b8"));
         btn_valider.setTextSize(15);
@@ -141,23 +142,23 @@ public class FragmentNewProduct {
         btn_valider.setFontIconSize(30);
 
 
-        btn_cancel = (MaterialFancyButton) dialogview.findViewById(R.id.btn_cancel);
+        btn_cancel = dialogview.findViewById(R.id.btn_cancel);
         btn_cancel.setBackgroundColor(Color.parseColor("#3498db"));
         btn_cancel.setFocusBackgroundColor(Color.parseColor("#5474b8"));
         btn_cancel.setTextSize(15);
         btn_cancel.setIconPosition(POSITION_LEFT);
         btn_cancel.setFontIconSize(30);
 
-        img_product = (ImageView) dialogview.findViewById(R.id.img_product);
+        img_product = dialogview.findViewById(R.id.img_product);
 
-        generate_codebarre = (ImageButton) dialogview.findViewById(R.id.generate_codebarre);
-        scan_codebarre = (ImageButton) dialogview.findViewById(R.id.scan_codebarre);
+        generate_codebarre = dialogview.findViewById(R.id.generate_codebarre);
+        scan_codebarre = dialogview.findViewById(R.id.scan_codebarre);
 
-        generate_reference = (ImageButton) dialogview.findViewById(R.id.generate_reference);
-        scan_reference = (ImageButton) dialogview.findViewById(R.id.scan_reference);
+        generate_reference = dialogview.findViewById(R.id.generate_reference);
+        scan_reference = dialogview.findViewById(R.id.scan_reference);
 
-        btn_from_gallery = (Button) dialogview.findViewById(R.id.btn_select_from_gallery);
-        btn_from_camera = (Button) dialogview.findViewById(R.id.btn_select_from_camera);
+        btn_from_gallery = dialogview.findViewById(R.id.btn_select_from_gallery);
+        btn_from_camera = dialogview.findViewById(R.id.btn_select_from_camera);
 
         ly_prix_achat = dialogview.findViewById(R.id.ly_prix_achat);
 
@@ -214,32 +215,48 @@ public class FragmentNewProduct {
         edt_prix5_ttc = dialogview.findViewById(R.id.edt_prix5_ttc);
         edt_prix6_ttc = dialogview.findViewById(R.id.edt_prix6_ttc);
 
-        //radioGroup_mode_tarif = (RadioGroup) dialogview.findViewById(R.id.rd_mode_tarif);
+        if(SOURCE_ACTIVITY.equals("EDIT_PRODUCT")){
 
-      /*  if(SOURCE_ACTIVITY.equals("EDIT_PRODUIT")){
+            edt_codebarre.setText(old_product.code_barre);
+            edt_codebarre.setEnabled(false);
+            edt_reference.setText(old_product.ref_produit);
+            edt_reference.setEnabled(false);
 
-            edt_designation.setText(old_client.client);
-            edt_codebarre.setText(old_client.adresse);
-            edt_prix_achat_ht.setText(old_client.tel);
-            edt_tva.setText(old_client.rc);
-            edt_prix_achat_ttc.setText(old_client.ifiscal);
-            edt_prix1_ht.setText(old_client.nis);
-            edt_prix2_ht.setText(old_client.ai);
-            edt_prix3_ht.setText(old_client.ai);
+            edt_designation.setText(old_product.produit);
+            edt_colissage.setText(nq.format(old_product.colissage));
 
-            if(old_client.mode_tarif.equals("0")){
-                radioGroup_mode_tarif.check(R.id.rb_1);
-            }
-            if(old_client.mode_tarif.equals("1")){
-                radioGroup_mode_tarif.check(R.id.rb_2);
-            }
-            if(old_client.mode_tarif.equals("2")){
-                radioGroup_mode_tarif.check(R.id.rb_3);
-            }
-            if(old_client.mode_tarif.equals("3")){
-                radioGroup_mode_tarif.check(R.id.rb_4);
-            }
-        }*/
+            edt_prix_achat_ht.setText(nf.format(old_product.pa_ht));
+            edt_tva.setText(nf.format(old_product.tva));
+
+            old_product.pa_ttc =  old_product.pa_ht + (old_product.pa_ht * old_product.tva / 100);
+            edt_prix_achat_ttc.setText(nf.format(old_product.pa_ttc));
+
+
+            edt_prix1_ht.setText(nf.format(old_product.pv1_ht));
+            edt_prix2_ht.setText(nf.format(old_product.pv2_ht));
+            edt_prix3_ht.setText(nf.format(old_product.pv3_ht));
+            edt_prix4_ht.setText(nf.format(old_product.pv4_ht));
+            edt_prix5_ht.setText(nf.format(old_product.pv5_ht));
+            edt_prix6_ht.setText(nf.format(old_product.pv6_ht));
+
+
+            old_product.pv1_ttc =  old_product.pv1_ht + (old_product.pv1_ht * old_product.tva / 100);
+            old_product.pv2_ttc =  old_product.pv2_ht + (old_product.pv2_ht * old_product.tva / 100);
+            old_product.pv3_ttc =  old_product.pv3_ht + (old_product.pv3_ht * old_product.tva / 100);
+            old_product.pv4_ttc =  old_product.pv4_ht + (old_product.pv4_ht * old_product.tva / 100);
+            old_product.pv5_ttc =  old_product.pv5_ht + (old_product.pv5_ht * old_product.tva / 100);
+            old_product.pv6_ttc =  old_product.pv6_ht + (old_product.pv6_ht * old_product.tva / 100);
+            edt_prix1_ttc.setText(nf.format(old_product.pv1_ttc));
+            edt_prix2_ttc.setText(nf.format(old_product.pv2_ttc));
+            edt_prix3_ttc.setText(nf.format(old_product.pv3_ttc));
+            edt_prix4_ttc.setText(nf.format(old_product.pv4_ttc));
+            edt_prix5_ttc.setText(nf.format(old_product.pv5_ttc));
+            edt_prix6_ttc.setText(nf.format(old_product.pv6_ttc));
+
+        }
+
+        ///////////////////
+        prefs = activity.getSharedPreferences(PREFS, MODE_PRIVATE);
 
         params = new PostData_Params();
         params = controller.select_params_from_database("SELECT * FROM PARAMS");
@@ -259,13 +276,13 @@ public class FragmentNewProduct {
         txt_input_prix6_ttc.setHint(params.pv6_titre + " (TTC)");
 
 
-        if(params.prix_2 == 1){
+        if(params.prix_2 == 1 || prefs.getBoolean("APP_AUTONOME", true)){
             lnr_prix2.setVisibility(View.VISIBLE);
         }else{
             lnr_prix2.setVisibility(View.INVISIBLE);
         }
 
-        if(params.prix_3 == 1){
+        if(params.prix_3 == 1 || prefs.getBoolean("APP_AUTONOME", true)){
             lnr_prix3.setVisibility(View.VISIBLE);
         }else{
             lnr_prix3.setVisibility(View.INVISIBLE);
@@ -535,19 +552,19 @@ public class FragmentNewProduct {
                 created_produit.isNew =  1;
 
                 created_produit.pv1_ht =  Double.parseDouble(edt_prix1_ht.getText().toString());
-                created_produit.pv1_ttc =  Double.parseDouble(edt_prix1_ttc.getText().toString());
+                created_produit.pv1_ttc =  created_produit.pv1_ht + (created_produit.pv1_ht * created_produit.tva / 100);
 
-                if(params.prix_2 == 1){
+                if(params.prix_2 == 1 || prefs.getBoolean("APP_AUTONOME", true)){
                     created_produit.pv2_ht =  Double.parseDouble(edt_prix2_ht.getText().toString());
-                    created_produit.pv2_ttc =  Double.parseDouble(edt_prix2_ttc.getText().toString());
+                    created_produit.pv2_ttc =  created_produit.pv2_ht + (created_produit.pv2_ht * created_produit.tva / 100);
                 }else {
                     created_produit.pv2_ht = 0.00;
                     created_produit.pv2_ttc = 0.00;
                 }
 
-                if(params.prix_3 == 1){
+                if(params.prix_3 == 1 || prefs.getBoolean("APP_AUTONOME", true)){
                     created_produit.pv3_ht =  Double.parseDouble(edt_prix3_ht.getText().toString());
-                    created_produit.pv3_ttc =  Double.parseDouble(edt_prix3_ttc.getText().toString());
+                    created_produit.pv3_ttc =  created_produit.pv3_ht + (created_produit.pv3_ht * created_produit.tva / 100);
                 }else {
                     created_produit.pv3_ht = 0.00;
                     created_produit.pv3_ttc = 0.00;
@@ -555,7 +572,7 @@ public class FragmentNewProduct {
 
                 if(params.prix_4 == 1){
                     created_produit.pv4_ht =  Double.parseDouble(edt_prix4_ht.getText().toString());
-                    created_produit.pv4_ttc =  Double.parseDouble(edt_prix4_ttc.getText().toString());
+                    created_produit.pv4_ttc = created_produit.pv4_ht + (created_produit.pv4_ht * created_produit.tva / 100);
                 }else {
                     created_produit.pv4_ht = 0.00;
                     created_produit.pv4_ttc = 0.00;
@@ -563,7 +580,7 @@ public class FragmentNewProduct {
 
                 if(params.prix_5 == 1){
                     created_produit.pv5_ht =  Double.parseDouble(edt_prix5_ht.getText().toString());
-                    created_produit.pv5_ttc =  Double.parseDouble(edt_prix5_ttc.getText().toString());
+                    created_produit.pv5_ttc =  created_produit.pv5_ht + (created_produit.pv5_ht * created_produit.tva / 100);
                 }else {
                     created_produit.pv5_ht = 0.00;
                     created_produit.pv5_ttc = 0.00;
@@ -571,35 +588,32 @@ public class FragmentNewProduct {
 
                 if(params.prix_6 == 1){
                     created_produit.pv6_ht =  Double.parseDouble(edt_prix6_ht.getText().toString());
-                    created_produit.pv6_ttc =  Double.parseDouble(edt_prix6_ttc.getText().toString());
+                    created_produit.pv6_ttc =  created_produit.pv6_ht + (created_produit.pv6_ht * created_produit.tva / 100);
                 }else {
                     created_produit.pv6_ht = 0.00;
                     created_produit.pv6_ttc = 0.00;
                 }
 
-                if(SOURCE_ACTIVITY.equals("EDIT_PRODUIT")){
+                if(SOURCE_ACTIVITY.equals("EDIT_PRODUCT")){
 
-                    //created_client.code_client = old_client.code_client;
-
-                    //Insert client into database,
-                    //boolean state_insert_client = controller.update_client(created_client);
-                   /* if(state_insert_client){
+                    //update client into database,
+                    boolean state_update_produit = controller.update_into_produit(created_produit);
+                    if(state_update_produit){
 
                         Crouton.makeText(activity, "Produit bien modifier", Style.INFO).show();
-
-                        SelectedClientEvent added_client = new SelectedClientEvent(created_client);
-                        bus.post(added_client);
+                        ProductEvent added_product_event = new ProductEvent(created_produit);
+                        bus.post(added_product_event);
 
                         dialog.dismiss();
 
                     }else{
-                        Crouton.makeText(activity, "Problème de mise à jour produit", Style.ALERT).show();
-                    }*/
+                        Crouton.makeText(activity, "Problème mise à jour produit", Style.ALERT).show();
+                    }
                 }else {
 
                     //update client into database,
-                    boolean state_update_produit = controller.insert_into_produit(created_produit);
-                    if(state_update_produit){
+                    boolean state_insert_produit = controller.insert_into_produit(created_produit);
+                    if(state_insert_produit){
 
                         Crouton.makeText(activity, "Produit bien ajouté", Style.INFO).show();
                         ProductEvent added_product_event = new ProductEvent(created_produit);
@@ -611,6 +625,9 @@ public class FragmentNewProduct {
                         Crouton.makeText(activity, "Problème insertion", Style.ALERT).show();
                     }
                 }
+
+                EventBus.getDefault().unregister(this);
+
             }
 
         });
